+++ pkg/osquery/table/platform_tables_linux.go
@@ -6,8 +6,10 @@ package table
 import (
 	"github.com/go-kit/kit/log"
 	"github.com/kolide/launcher/pkg/osquery/tables/crowdstrike/falcon_kernel_check"
+	"github.com/kolide/launcher/pkg/osquery/tables/crowdstrike/falconctl"
 	"github.com/kolide/launcher/pkg/osquery/tables/cryptsetup"
 	"github.com/kolide/launcher/pkg/osquery/tables/dataflattentable"
+	"github.com/kolide/launcher/pkg/osquery/tables/execparsers/simple_array"
 	"github.com/kolide/launcher/pkg/osquery/tables/fscrypt_info"
 	"github.com/kolide/launcher/pkg/osquery/tables/gsettings"
 	"github.com/kolide/launcher/pkg/osquery/tables/secureboot"
@@ -24,6 +26,8 @@ func platformTables(client *osquery.ExtensionManagerClient, logger log.Logger, c
 		xrdb.TablePlugin(client, logger),
 		fscrypt_info.TablePlugin(logger),
 		falcon_kernel_check.TablePlugin(logger),
+		falconctl.NewFalconctlOptionTable(logger),
+
 		dataflattentable.TablePluginExec(client, logger,
 			"kolide_nmcli_wifi", dataflattentable.KeyValueType,
 			[]string{"/usr/bin/nmcli", "--mode=multiline", "--fields=all", "device", "wifi", "list"},
@@ -32,5 +36,6 @@ func platformTables(client *osquery.ExtensionManagerClient, logger log.Logger, c
 			[]string{"lsblk", "-J"},
 			dataflattentable.WithBinDirs("/usr/bin", "/bin"),
 		),
+		dataflattentable.NewExecAndParseTable(logger, "kolide_falconctl_systags", simple_array.Parser, []string{"/opt/CrowdStrike/falconctl", "-g", "--systags"}),
 	}
 }
