+++ pkg/osquery/tables/execparsers/simple_array/parser.go
@@ -1,4 +1,4 @@
-package falconctl
+package simple_array
 
 import (
 	"bufio"
@@ -7,19 +7,33 @@ import (
 	"strings"
 )
 
-// parseSystags parses the systags. As we understand it, this is a simple array of strings. One per line. Maybe comma delimited.
-func parseSystags(reader io.Reader) (any, error) {
+type parser struct {
+}
+
+func (p parser) Parse(reader io.Reader) (any, error) {
+	return parse(reader)
+}
+
+func New() parser {
+	return parser{}
+}
+
+// Parser is a parser for dsregcmd output
+var Parser = New()
+
+// parser parses line by line, additionally splitting on commas. As this is an array of simple strings, anything with a space is an error
+func parse(reader io.Reader) (any, error) {
 	found := make([]string, 0)
 
 	scanner := bufio.NewScanner(reader)
 	for scanner.Scan() {
 		line := scanner.Text()
-		line = strings.TrimSpace(line)
 
 		for _, chunk := range strings.Split(line, ",") {
 
 			// trim quotes and spaces
-			chunk = strings.TrimSpace(strings.Trim(chunk, `"`))
+			chunk = strings.TrimSpace(chunk)
+			chunk = strings.Trim(chunk, `"`)
 
 			// If a chunk has a space in the middle, it's malformed and we should error out
 			if strings.Contains(chunk, " ") {
@@ -28,7 +42,6 @@ func parseSystags(reader io.Reader) (any, error) {
 
 			found = append(found, chunk)
 		}
-
 	}
 
 	return found, nil
