+++ pkg/osquery/tables/execparsers/simple_array/test-data/complex.txt
@@ -1,0 +1,5 @@
+one
+two,  three
+four
+five,six,seven
+"eight"
