+++ pkg/osquery/tables/execparsers/simple_array/test-data/simple.txt
